cinechunks-web/requirements.txt
fastapi
uvicorn[standard]
jinja2
openai
fastmcp
python-dotenv
